add flake.nix
The most important change was that now builtins.currentSystem does not
exist anymore. This means we need to have one system and import
nixpkgs with it and release does not contain all systems anymore. I
have no idea about the implications here, I chose the easiest way IMO
to get the ball rolling, but I'm happy to go back to the drawing board
if needed.